back/requirements.txt
asgiref==3.10.0
Django==4.2.15
django-cors-headers==4.9.0
django-dotenv==1.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
celery==5.4.0
redis==5.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
